check if command exist before aliasing

+++ bash_aliases
@@ -40,8 +40,7 @@ Unpack ()
   fi
 }
 
-
-#for some configurations run mc in ASCII mode
+#for some cases better to run mc in ASCII mode
 #alias mc="mc -a"
 
 #Systemd/systemctl
@@ -71,10 +70,12 @@ then
     alias Purge="sudo apt-get remove –purge"
 else
     # Centos
+    # for Yum
     #alias Install="sudo yum install"
     #alias Update="sudo yum update"
     #alias Upgrade="sudo yum upgrade"
     #alias Search="yum  search"
+    # for dnf
     alias Install="sudo dnf install"
     alias Update="sudo dnf update"
     alias Upgrade="sudo dnf upgrade"
@@ -89,7 +90,6 @@ fi
 
 #File command aliases
 alias Ls="ls -p --color=auto --group-directories-first"
-alias Hidden="! git ls-files -v . | grep ^S | awk '{ print $2 }'" 
 alias Top10files="find . -type f -exec ls -sh {} \; | sort -n -r | head -10"
 #copy using rsync
 alias Rsync="rsync -avv --stats --human-readable --itemize-changes --progress --partial"
@@ -107,7 +107,6 @@ alias Ping='ping -c 5'
 #Get HTTP Headers
 alias Headers="curl -I"
 
-
 #System monitoring aliases
 alias Mhz='watch -n 1 "cat /proc/cpuinfo | grep MHz"'
 alias Mem="free -m -l -t"
@@ -121,39 +120,46 @@ alias Psmem="ps auxf | sort -nr -k 4"
 alias Psmem10="ps auxf | sort -nr -k 4 | head -10"
 
 # Git Aliases
-alias Gstat="git status"
-alias Gbra="git branch"
-alias Gcheck="git checkout"
-alias Glog="git log --oneline --decorate --color"
-alias Gla="git for-each-ref --sort=committerdate refs/heads/ --format='%(HEAD) %(color:yellow)%(refname:short)%(color:reset) - %(color:red)%(objectname:short)%(color:reset) - %(contents:subject) - %(authorname) - (%(color:green)%(committerdate:relative)%(color:reset))'"
-alias Gpu="git push --set-upstream origin"
-alias Gamend="git commit --amend"
-alias Gcommit="git commit -m"
-alias Gnuke="git clean -df && git reset --hard"
-alias Gpop="git stash pop"
-alias Gpull="git pull"
-alias Gpush="git push"
-alias Gstash="git stash"
-alias Gadd="git add"
-alias Gdiff="git diff"
+which git 2>/dev/null 1>/dev/null
+if [ $? -eq 0 ]; then
+  alias Gstat="git status"
+  alias Gbra="git branch"
+  alias Gcheck="git checkout"
+  alias Glog="git log --oneline --decorate --color"
+  alias Gla="git for-each-ref --sort=committerdate refs/heads/ --format='%(HEAD) %(color:yellow)%(refname:short)%(color:reset) - %(color:red)%(objectname:short)%(color:reset) - %(contents:subject) - %(authorname) - (%(color:green)%(committerdate:relative)%(color:reset))'"
+  alias Gpu="git push --set-upstream origin"
+  alias Gamend="git commit --amend"
+  alias Gcommit="git commit -m"
+  alias Gnuke="git clean -df && git reset --hard"
+  alias Gpop="git stash pop"
+  alias Gpull="git pull"
+  alias Gpush="git push"
+  alias Gstash="git stash"
+  alias Gadd="git add"
+  alias Gdiff="git diff"
+  alias Ghidden="! git ls-files -v . | grep ^S | awk '{ print $2 }'" 
+fi
 
 #Docker aliases
-alias Deit="docker exec -it"
-alias Di="docker images"
-alias Dil="docker images | sed -n '2p' | awk '{print $3}' | pbcopy"
-alias Dk="docker kill"
-alias Dkl="docker ps | sed -n '2p' | awk '{print $1}' | xargs docker kill"
-alias Dps="docker ps"
-alias Dpsl="docker ps | sed -n '2p' | awk '{print $1}'"
-alias Dritrm="docker run -it --rm"
-alias Drmid="docker rmi $(docker images --filter 'dangling=true' -q --no-trunc)"
-alias Drmif="docker rmi -f"
+which docker 2>/dev/null 1>/dev/null
+if [ $? -eq 0 ]; then
+  alias Deit="docker exec -it"
+  alias Di="docker images"
+  alias Dil="docker images | sed -n '2p' | awk '{print $3}' | pbcopy"
+  alias Dk="docker kill"
+  alias Dkl="docker ps | sed -n '2p' | awk '{print $1}' | xargs docker kill"
+  alias Dps="docker ps"
+  alias Dpsl="docker ps | sed -n '2p' | awk '{print $1}'"
+  alias Dritrm="docker run -it --rm"
+  alias Drmid="docker rmi $(docker images --filter 'dangling=true' -q --no-trunc)"
+  alias Drmif="docker rmi -f"
+fi
 
 # Command line head / tail shortcuts
-alias H='| head'
-alias T='| tail'
-alias G='| grep'
-alias L="| less"
+alias H='head'
+alias T='tail'
+alias G='grep'
+alias L="less"
 
 #history search
 alias History="history | grep"
